Add files via upload
This jupyter notebook is where I perform one of my Portfolio Project, from Codecademy DataScientist: Machine Learning career path. 
It is a learning Project which allows to fail and learn from that. 
This is my first work with jupyter notebooks at this level and I know that I have a long learning journey to live.

Thank you for your support in advance.